changed filterSearch func + small changes
